Update and unneeded line removal
I removed the start button and changed it to a Show/Hide button so if you want to have passwords you can without others seeing them. Please note the save file is not encrypted. That might be next update now that I can actually make changes to the software again.